refactor(ts): split server.ts into tool modules
server.ts (1173 lines) split into: server.ts entry (Express + session mgmt, 201 lines), tools/prtsTools (6 wiki tools), tools/gamedataTools (16 game-data tools), tools/storyTools (8 story tools + requireStoryZip), startupSync (sync orchestration, 204 lines).

toolSurface test updated to scan src/tools/ dir and compare as sets (order not frozen). typecheck clean, 138 tests pass.
